🔨 Move `actions.yaml` related types to `actions.yaml.ts`
